+++ src/main/java/com/codeprometheus/aialgohelper/plugin/window/navigator/NavigatorPanel.java
@@ -254,11 +254,6 @@ public class NavigatorPanel extends SimpleToolWindowPanel implements NavigatorPa
     }
 
     @Override
-    public Object getData(String dataId) {
-        return super.getData(dataId);
-    }
-
-    @Override
     public void dispose() {
         navigatorTable.dispose();
     }